Create AccountRating.vfp
Objective: Create a Vf page that allows users to edit an Account's Rating field value, and to see a table of all the open- and closed-won child Opportunities (to help users decide on a rating).  The table will have four columns: Name, Amount, Stage, and Close Date.  If the user submits the Rating value "Hot," create a Task on all open child Opportunities.  Next, assign the Task to the Opportunity owner, sets its Status field value to "Not Started," and set its Subject field value to "Send follow-up email to primary contact."